requirements.txt
streamlit
pandas
numpy
ckanapi
python-dotenv
groq
